+++ com.unity.render-pipelines.high-definition/Editor/ShaderGraph/HDKeywords.cs
@@ -6,6 +6,16 @@ namespace UnityEditor.Rendering.HighDefinition.ShaderGraph
     {
         public static class Descriptors
         {
+
+            public static KeywordDescriptor WriteNormalBufferDefine = new KeywordDescriptor()
+            {
+                displayName = "Write Normal Buffer",
+                referenceName = "WRITE_NORMAL_BUFFER",
+                type = KeywordType.Boolean,
+                definition = KeywordDefinition.Predefined,
+                scope = KeywordScope.Global,
+            };
+            
             public static KeywordDescriptor WriteNormalBuffer = new KeywordDescriptor()
             {
                 displayName = "Write Normal Buffer",
@@ -269,7 +279,7 @@ namespace UnityEditor.Rendering.HighDefinition.ShaderGraph
         {
             { HDBase },
             { Descriptors.AlphaTest, new FieldCondition(HDFields.AlphaTestPrepass, true) },
-            { Descriptors.WriteNormalBuffer, new FieldCondition(HDFields.DisableSSRTransparent, true) },
+            { Descriptors.WriteNormalBufferDefine, new FieldCondition(HDFields.DisableSSRTransparent, false) },
         };
 
         public static KeywordCollection TransparentDepthPostpass = new KeywordCollection